# packages/contracts/openapi/rc-racing-v1.yaml
openapi: 3.1.0
info:
  title: RC Racing API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
...
  /v1/ride-offers/{id}/accept:
    post:
      operationId: acceptRideOffer
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201":
          description: Accepted ride snapshot
  /v1/rides/{id}:
    get:
      operationId: getRide
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Authoritative ride snapshot
...